backend/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.30
pymysql==1.1.1
aiomysql==0.2.0

# AI / LangChain
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
chromadb==0.5.0
langchain-chroma==0.1.2
python-pptx==0.6.23

# OCR
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
pypdf==4.2.0

# ML / Image
opencv-python-headless==4.10.0.84

# PDF Generation
reportlab==4.2.0

# Translation
deep-translator==1.11.4

# Utilities
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
websockets==12.0
gtts==2.5.1
openai==1.30.0
groq==0.9.0
